requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-cities-light==3.9.2
django-ckeditor==6.5.1
django-debug-toolbar==4.2.0
django-htmx==1.16.0
django-js-asset==2.0.0
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
fontawesomefree==6.4.0
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.15
Pillow==9.5.0
progressbar2==4.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python-utils==3.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
Unidecode==1.3.6
urllib3==2.0.3
